pwd + grp + resource: extract into module/{pwd,grp,resource}/
pwd carries its uid_converter helper.  importing.rs shrinks by 707 lines.
