+++ pyre/pyre-interpreter/src/importing.rs
@@ -132,11 +132,11 @@ pub fn install_builtin_modules() {
     register_builtin_module("_signal", crate::module::_signal::moduledef::init);
     register_builtin_module("atexit", crate::module::atexit::moduledef::init);
     #[cfg(unix)]
-    register_builtin_module("pwd", init_pwd);
+    register_builtin_module("pwd", crate::module::pwd::moduledef::init);
     #[cfg(unix)]
-    register_builtin_module("grp", init_grp);
+    register_builtin_module("grp", crate::module::grp::moduledef::init);
     #[cfg(unix)]
-    register_builtin_module("resource", init_resource);
+    register_builtin_module("resource", crate::module::resource::moduledef::init);
     #[cfg(unix)]
     register_builtin_module("fcntl", crate::module::fcntl::moduledef::init);
     #[cfg(unix)]
@@ -1210,717 +1210,6 @@ fn stat_result_type() -> PyObjectRef {
             tp
         })
     })
-}
-
-/// `interp_pwd.py:50-73 uid_converter` — narrow a python int to `uid_t`.
-///
-/// `-1` is the "current uid" sentinel and passes through unchanged
-/// (cast to `uid_t` it becomes the max value, matching the C convention
-/// most BSDs use).  Other negative inputs raise OverflowError "user id
-/// is less than minimum"; values that don't fit in `uid_t` raise
-/// OverflowError "user id is greater than maximum".  Floats / non-int
-/// inputs raise TypeError via `int_w`.
-#[cfg(unix)]
-fn pwd_uid_converter(w_uid: pyre_object::PyObjectRef) -> Result<libc::uid_t, crate::PyError> {
-    let val = match crate::baseobjspace::int_w(w_uid) {
-        Ok(v) => v,
-        Err(e) if matches!(e.kind, crate::PyErrorKind::OverflowError) => {
-            // `interp_pwd.py:60-66` — fall through to `uint_w` and
-            // map to "greater than maximum" / "less than minimum"
-            // OverflowError.  pyre's `int_w` only surfaces the
-            // positive overflow case; the negative bigint path is
-            // unreachable here because a negative bigint that
-            // fails to fit i64 is more-negative than i64::MIN —
-            // user id less than minimum.
-            return Err(crate::PyError::overflow_error(
-                "user id is greater than maximum",
-            ));
-        }
-        Err(e) => return Err(e),
-    };
-    if val == -1 {
-        return Ok((-1i64) as libc::uid_t);
-    }
-    if val < 0 {
-        return Err(crate::PyError::overflow_error(
-            "user id is less than minimum",
-        ));
-    }
-    let uid = val as libc::uid_t;
-    if uid as i64 != val {
-        return Err(crate::PyError::overflow_error(
-            "user id is greater than maximum",
-        ));
-    }
-    Ok(uid)
-}
-
-/// pwd module — `pypy/module/pwd/interp_pwd.py`.
-///
-/// getpwuid / getpwnam / getpwall return 7-tuples with the
-/// `(pw_name, pw_passwd, pw_uid, pw_gid, pw_gecos, pw_dir, pw_shell)`
-/// layout.  `struct_passwd` / `struct_pwent` are exposed as the same
-/// builtin type so `isinstance(pwd.struct_passwd, type)` succeeds and
-/// `pwd.struct_passwd` is identity-equal to `pwd.struct_pwent`
-/// (`app_pwd.py:1-21`).  Full structseq instance materialisation
-/// (so `pw_entry.pw_name` returns a string) is a framework prereq
-/// tracked separately.
-///
-/// Backed by `rustpython_host_env::pwd` (a thin `nix` wrapper).
-#[cfg(unix)]
-fn init_pwd(ns: &mut DictStorage) {
-    #[cfg(feature = "host_env")]
-    fn make_struct_passwd(pw: &rustpython_host_env::pwd::Passwd) -> pyre_object::PyObjectRef {
-        pyre_object::w_tuple_new(vec![
-            pyre_object::w_str_new(&pw.name),
-            pyre_object::w_str_new(&pw.passwd),
-            pyre_object::w_int_new(pw.uid as i64),
-            pyre_object::w_int_new(pw.gid as i64),
-            pyre_object::w_str_new(&pw.gecos),
-            pyre_object::w_str_new(&pw.dir),
-            pyre_object::w_str_new(&pw.shell),
-        ])
-    }
-    // `interp_pwd.py:75-87 make_struct_passwd` libc backend, used when
-    // the host_env abstraction layer is disabled.  Mirrors the same
-    // rffi.charp2str / int construction PyPy uses.
-    #[cfg(not(feature = "host_env"))]
-    unsafe fn make_struct_passwd_libc(pw: *const libc::passwd) -> pyre_object::PyObjectRef {
-        unsafe fn cstr(p: *const libc::c_char) -> String {
-            if p.is_null() {
-                String::new()
-            } else {
-                std::ffi::CStr::from_ptr(p).to_string_lossy().into_owned()
-            }
-        }
-        pyre_object::w_tuple_new(vec![
-            pyre_object::w_str_new(&cstr((*pw).pw_name)),
-            pyre_object::w_str_new(&cstr((*pw).pw_passwd)),
-            pyre_object::w_int_new((*pw).pw_uid as i64),
-            pyre_object::w_int_new((*pw).pw_gid as i64),
-            pyre_object::w_str_new(&cstr((*pw).pw_gecos)),
-            pyre_object::w_str_new(&cstr((*pw).pw_dir)),
-            pyre_object::w_str_new(&cstr((*pw).pw_shell)),
-        ])
-    }
-    // `app_pwd.py:1-21` — `pwd.struct_passwd` / `pwd.struct_pwent`.
-    let struct_passwd_type = crate::typedef::make_builtin_type("pwd.struct_passwd", |_| {});
-    crate::dict_storage_store(ns, "struct_passwd", struct_passwd_type);
-    crate::dict_storage_store(ns, "struct_pwent", struct_passwd_type);
-    crate::dict_storage_store(
-        ns,
-        "getpwuid",
-        crate::make_builtin_function_with_arity(
-            "getpwuid",
-            |args| {
-                if args.is_empty() {
-                    return Err(crate::PyError::type_error("getpwuid() missing argument"));
-                }
-                // `interp_pwd.py:50-73 uid_converter`: -1 sentinel passes
-                // through; negative-other → OverflowError "less than
-                // minimum"; positive-too-big → OverflowError "greater
-                // than maximum".  `interp_pwd.py:97-100 getpwuid` catches
-                // OverflowError and converts it to KeyError "uid not
-                // found".
-                let uid = match pwd_uid_converter(args[0]) {
-                    Ok(u) => u,
-                    Err(e) if matches!(e.kind, crate::PyErrorKind::OverflowError) => {
-                        return Err(crate::PyError::key_error("getpwuid(): uid not found"));
-                    }
-                    Err(e) => return Err(e),
-                };
-                #[cfg(feature = "host_env")]
-                {
-                    match rustpython_host_env::pwd::getpwuid(uid) {
-                        Ok(Some(pw)) => return Ok(make_struct_passwd(&pw)),
-                        Ok(None) => {
-                            return Err(crate::PyError::key_error(format!(
-                                "getpwuid(): uid not found: {}",
-                                uid as i64
-                            )));
-                        }
-                        Err(e) => {
-                            return Err(crate::PyError::os_error_with_errno(
-                                e.raw_os_error().unwrap_or(0),
-                                format!("getpwuid: {e}"),
-                            ));
-                        }
-                    }
-                }
-                // `interp_pwd.py:90-108` — libc fallback path; host_env
-                // is a pyre-only abstraction layer over the same
-                // getpwuid() call PyPy makes via rffi.llexternal.
-                #[cfg(not(feature = "host_env"))]
-                unsafe {
-                    let pw = libc::getpwuid(uid);
-                    if pw.is_null() {
-                        return Err(crate::PyError::key_error(format!(
-                            "getpwuid(): uid not found: {}",
-                            uid as i64
-                        )));
-                    }
-                    return Ok(make_struct_passwd_libc(pw));
-                }
-            },
-            1,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "getpwnam",
-        crate::make_builtin_function_with_arity(
-            "getpwnam",
-            |args| {
-                if args.is_empty() {
-                    return Err(crate::PyError::type_error("getpwnam() missing argument"));
-                }
-                if !unsafe { pyre_object::is_str(args[0]) } {
-                    return Err(crate::PyError::type_error(
-                        "getpwnam(): name should be a string",
-                    ));
-                }
-                let name = unsafe { pyre_object::w_str_get_value(args[0]) };
-                // `interp_pwd.py:111 @unwrap_spec(name='text0')` rejects
-                // embedded NULs.  CString::new() enforces that here.
-                let c_name = std::ffi::CString::new(name).map_err(|_| {
-                    crate::PyError::value_error("getpwnam: name must not contain NUL bytes")
-                })?;
-                #[cfg(feature = "host_env")]
-                {
-                    match rustpython_host_env::pwd::getpwnam(name) {
-                        Some(pw) => return Ok(make_struct_passwd(&pw)),
-                        None => {
-                            return Err(crate::PyError::key_error(format!(
-                                "getpwnam(): name not found: {}",
-                                name
-                            )));
-                        }
-                    }
-                }
-                #[cfg(not(feature = "host_env"))]
-                unsafe {
-                    let pw = libc::getpwnam(c_name.as_ptr());
-                    if pw.is_null() {
-                        return Err(crate::PyError::key_error(format!(
-                            "getpwnam(): name not found: {}",
-                            name
-                        )));
-                    }
-                    return Ok(make_struct_passwd_libc(pw));
-                }
-            },
-            1,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "getpwall",
-        crate::make_builtin_function_with_arity(
-            "getpwall",
-            |_| {
-                #[cfg(feature = "host_env")]
-                {
-                    let items: Vec<pyre_object::PyObjectRef> = rustpython_host_env::pwd::getpwall()
-                        .iter()
-                        .map(make_struct_passwd)
-                        .collect();
-                    return Ok(pyre_object::w_list_new(items));
-                }
-                // `interp_pwd.py:123-134` — setpwent / loop getpwent /
-                // endpwent.
-                #[cfg(not(feature = "host_env"))]
-                unsafe {
-                    let mut items: Vec<pyre_object::PyObjectRef> = Vec::new();
-                    libc::setpwent();
-                    loop {
-                        let pw = libc::getpwent();
-                        if pw.is_null() {
-                            break;
-                        }
-                        items.push(make_struct_passwd_libc(pw));
-                    }
-                    libc::endpwent();
-                    return Ok(pyre_object::w_list_new(items));
-                }
-            },
-            0,
-        ),
-    );
-}
-
-/// grp module — `lib_pypy/grp.py` (PyPy keeps it app-level via
-/// `_pwdgrp_cffi`).  pyre takes CPython's `Modules/grpmodule.c`
-/// shape since pyre has no app-level stdlib.
-///
-/// getgrgid / getgrnam / getgrall return 4-tuples `(gr_name,
-/// gr_passwd, gr_gid, gr_mem)` matching CPython.  `grp.struct_group`
-/// is exposed as a builtin type attribute; full structseq instance
-/// materialisation (so `entry.gr_name` works) is blocked on the
-/// structseq framework task.
-#[cfg(unix)]
-fn init_grp(ns: &mut DictStorage) {
-    #[cfg(feature = "host_env")]
-    fn make_struct_group(g: &rustpython_host_env::grp::Group) -> pyre_object::PyObjectRef {
-        let mem_items: Vec<pyre_object::PyObjectRef> =
-            g.mem.iter().map(|s| pyre_object::w_str_new(s)).collect();
-        pyre_object::w_tuple_new(vec![
-            pyre_object::w_str_new(&g.name),
-            pyre_object::w_str_new(&g.passwd),
-            pyre_object::w_int_new(g.gid as i64),
-            pyre_object::w_list_new(mem_items),
-        ])
-    }
-    // `lib_pypy/grp.py:21-34 _group_from_gstruct` libc backend, used when
-    // the host_env abstraction layer is disabled.
-    #[cfg(not(feature = "host_env"))]
-    unsafe fn make_struct_group_libc(g: *const libc::group) -> pyre_object::PyObjectRef {
-        unsafe fn cstr(p: *const libc::c_char) -> String {
-            if p.is_null() {
-                String::new()
-            } else {
-                std::ffi::CStr::from_ptr(p).to_string_lossy().into_owned()
-            }
-        }
-        let mut mem_items: Vec<pyre_object::PyObjectRef> = Vec::new();
-        let mut p = (*g).gr_mem;
-        if !p.is_null() {
-            while !(*p).is_null() {
-                mem_items.push(pyre_object::w_str_new(&cstr(*p)));
-                p = p.add(1);
-            }
-        }
-        pyre_object::w_tuple_new(vec![
-            pyre_object::w_str_new(&cstr((*g).gr_name)),
-            pyre_object::w_str_new(&cstr((*g).gr_passwd)),
-            pyre_object::w_int_new((*g).gr_gid as i64),
-            pyre_object::w_list_new(mem_items),
-        ])
-    }
-    // `lib_pypy/grp.py:13-19 class struct_group` — exposed so
-    // `grp.struct_group` is observable on the module even though
-    // returned values are still raw tuples.
-    let struct_group_type = crate::typedef::make_builtin_type("grp.struct_group", |_| {});
-    crate::dict_storage_store(ns, "struct_group", struct_group_type);
-    crate::dict_storage_store(
-        ns,
-        "getgrgid",
-        crate::make_builtin_function_with_arity(
-            "getgrgid",
-            |args| {
-                if args.is_empty() {
-                    return Err(crate::PyError::type_error("getgrgid() missing argument"));
-                }
-                // `Modules/grpmodule.c grp_getgrgid` — accept any
-                // python int (including bigint) via int_w; reject
-                // floats as TypeError.  PyPy's `lib_pypy/grp.py`
-                // forwards directly through ctypes which would
-                // do the same conversion.
-                let val = crate::baseobjspace::int_w(args[0])?;
-                let gid = val as libc::gid_t;
-                #[cfg(feature = "host_env")]
-                {
-                    match rustpython_host_env::grp::getgrgid(gid) {
-                        Ok(Some(g)) => return Ok(make_struct_group(&g)),
-                        Ok(None) => {
-                            return Err(crate::PyError::key_error(format!(
-                                "getgrgid(): gid not found: {}",
-                                gid
-                            )));
-                        }
-                        Err(e) => {
-                            return Err(crate::PyError::os_error_with_errno(
-                                e.raw_os_error().unwrap_or(0),
-                                format!("getgrgid: {e}"),
-                            ));
-                        }
-                    }
-                }
-                #[cfg(not(feature = "host_env"))]
-                unsafe {
-                    let g = libc::getgrgid(gid);
-                    if g.is_null() {
-                        return Err(crate::PyError::key_error(format!(
-                            "getgrgid(): gid not found: {}",
-                            gid
-                        )));
-                    }
-                    return Ok(make_struct_group_libc(g));
-                }
-            },
-            1,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "getgrnam",
-        crate::make_builtin_function_with_arity(
-            "getgrnam",
-            |args| {
-                if args.is_empty() {
-                    return Err(crate::PyError::type_error("getgrnam() missing argument"));
-                }
-                if !unsafe { pyre_object::is_str(args[0]) } {
-                    return Err(crate::PyError::type_error(
-                        "getgrnam(): name should be a string",
-                    ));
-                }
-                let name = unsafe { pyre_object::w_str_get_value(args[0]) };
-                // Reject embedded NULs (parity with PyPy's @unwrap_spec
-                // text0 used for similar lookup APIs).
-                let c_name = std::ffi::CString::new(name).map_err(|_| {
-                    crate::PyError::value_error("getgrnam: name must not contain NUL bytes")
-                })?;
-                #[cfg(feature = "host_env")]
-                {
-                    match rustpython_host_env::grp::getgrnam(name) {
-                        Ok(Some(g)) => return Ok(make_struct_group(&g)),
-                        Ok(None) => {
-                            return Err(crate::PyError::key_error(format!(
-                                "getgrnam(): name not found: {}",
-                                name
-                            )));
-                        }
-                        Err(e) => {
-                            return Err(crate::PyError::os_error_with_errno(
-                                e.raw_os_error().unwrap_or(0),
-                                format!("getgrnam: {e}"),
-                            ));
-                        }
-                    }
-                }
-                #[cfg(not(feature = "host_env"))]
-                unsafe {
-                    let g = libc::getgrnam(c_name.as_ptr());
-                    if g.is_null() {
-                        return Err(crate::PyError::key_error(format!(
-                            "getgrnam(): name not found: {}",
-                            name
-                        )));
-                    }
-                    return Ok(make_struct_group_libc(g));
-                }
-            },
-            1,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "getgrall",
-        crate::make_builtin_function_with_arity(
-            "getgrall",
-            |_| {
-                #[cfg(feature = "host_env")]
-                {
-                    let items: Vec<pyre_object::PyObjectRef> = rustpython_host_env::grp::getgrall()
-                        .iter()
-                        .map(make_struct_group)
-                        .collect();
-                    return Ok(pyre_object::w_list_new(items));
-                }
-                #[cfg(not(feature = "host_env"))]
-                unsafe {
-                    let mut items: Vec<pyre_object::PyObjectRef> = Vec::new();
-                    libc::setgrent();
-                    loop {
-                        let g = libc::getgrent();
-                        if g.is_null() {
-                            break;
-                        }
-                        items.push(make_struct_group_libc(g));
-                    }
-                    libc::endgrent();
-                    return Ok(pyre_object::w_list_new(items));
-                }
-            },
-            0,
-        ),
-    );
-}
-
-/// resource module — `lib_pypy/resource.py` (PyPy keeps it app-level
-/// via `_resource_cffi`).  pyre takes CPython's `Modules/resource.c`
-/// shape since pyre has no app-level stdlib.
-///
-/// Exposes getrusage / getrlimit / setrlimit plus the standard RUSAGE_*
-/// and RLIMIT_* constants, the `struct_rusage` type attribute, and the
-/// `error = OSError` alias.  Backed by `rustpython_host_env::resource`.
-fn init_resource(ns: &mut DictStorage) {
-    // `lib_pypy/resource.py:13 error = OSError` and
-    // `:15-37 class struct_rusage`.
-    let w_os_error = crate::builtins::lookup_exc_class("OSError")
-        .expect("OSError must be installed before init_resource");
-    crate::dict_storage_store(ns, "error", w_os_error);
-    crate::dict_storage_store(
-        ns,
-        "struct_rusage",
-        crate::typedef::make_builtin_type("resource.struct_rusage", |_| {}),
-    );
-    // ── struct_rusage tuple (16-field layout matches CPython) ──
-    #[cfg(all(unix, feature = "host_env"))]
-    fn make_struct_rusage(r: &rustpython_host_env::resource::RUsage) -> pyre_object::PyObjectRef {
-        let tv_to_f = |tv: libc::timeval| tv.tv_sec as f64 + (tv.tv_usec as f64) * 1e-6;
-        pyre_object::w_tuple_new(vec![
-            pyre_object::floatobject::w_float_new(tv_to_f(r.ru_utime)),
-            pyre_object::floatobject::w_float_new(tv_to_f(r.ru_stime)),
-            pyre_object::w_int_new(r.ru_maxrss as i64),
-            pyre_object::w_int_new(r.ru_ixrss as i64),
-            pyre_object::w_int_new(r.ru_idrss as i64),
-            pyre_object::w_int_new(r.ru_isrss as i64),
-            pyre_object::w_int_new(r.ru_minflt as i64),
-            pyre_object::w_int_new(r.ru_majflt as i64),
-            pyre_object::w_int_new(r.ru_nswap as i64),
-            pyre_object::w_int_new(r.ru_inblock as i64),
-            pyre_object::w_int_new(r.ru_oublock as i64),
-            pyre_object::w_int_new(r.ru_msgsnd as i64),
-            pyre_object::w_int_new(r.ru_msgrcv as i64),
-            pyre_object::w_int_new(r.ru_nsignals as i64),
-            pyre_object::w_int_new(r.ru_nvcsw as i64),
-            pyre_object::w_int_new(r.ru_nivcsw as i64),
-        ])
-    }
-    crate::dict_storage_store(
-        ns,
-        "getrusage",
-        crate::make_builtin_function_with_arity(
-            "getrusage",
-            |args| {
-                #[cfg(all(unix, feature = "host_env"))]
-                {
-                    let who = if let Some(&a) = args.first() {
-                        if unsafe { pyre_object::is_int(a) } {
-                            unsafe { pyre_object::w_int_get_value(a) as i32 }
-                        } else {
-                            return Err(crate::PyError::type_error(
-                                "getrusage(): who should be an integer",
-                            ));
-                        }
-                    } else {
-                        return Err(crate::PyError::type_error("getrusage() missing argument"));
-                    };
-                    match rustpython_host_env::resource::getrusage(who) {
-                        Ok(r) => return Ok(make_struct_rusage(&r)),
-                        Err(e) => {
-                            let errno = e.raw_os_error().unwrap_or(0);
-                            // `lib_pypy/resource.py:106` raises ValueError for
-                            // an invalid `who`; only other errno values are
-                            // surfaced as OSError.
-                            if errno == libc::EINVAL {
-                                return Err(crate::PyError::value_error("invalid who parameter"));
-                            }
-                            return Err(crate::PyError::os_error_with_errno(
-                                errno,
-                                format!("getrusage: {e}"),
-                            ));
-                        }
-                    }
-                }
-                #[cfg(not(all(unix, feature = "host_env")))]
-                {
-                    let _ = args;
-                    Err(crate::PyError::not_implemented(
-                        "resource.getrusage requires host_env feature",
-                    ))
-                }
-            },
-            1,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "getrlimit",
-        crate::make_builtin_function_with_arity(
-            "getrlimit",
-            |args| {
-                #[cfg(all(unix, feature = "host_env"))]
-                {
-                    let res = if let Some(&a) = args.first() {
-                        if unsafe { pyre_object::is_int(a) } {
-                            unsafe { pyre_object::w_int_get_value(a) as libc::rlim_t }
-                        } else {
-                            return Err(crate::PyError::type_error(
-                                "getrlimit(): resource should be an integer",
-                            ));
-                        }
-                    } else {
-                        return Err(crate::PyError::type_error("getrlimit() missing argument"));
-                    };
-                    match rustpython_host_env::resource::getrlimit(res) {
-                        Ok(rl) => {
-                            return Ok(pyre_object::w_tuple_new(vec![
-                                pyre_object::w_int_new(rl.rlim_cur as i64),
-                                pyre_object::w_int_new(rl.rlim_max as i64),
-                            ]));
-                        }
-                        Err(e) => {
-                            return Err(crate::PyError::os_error_with_errno(
-                                e.raw_os_error().unwrap_or(0),
-                                format!("getrlimit: {e}"),
-                            ));
-                        }
-                    }
-                }
-                #[cfg(not(all(unix, feature = "host_env")))]
-                {
-                    let _ = args;
-                    Err(crate::PyError::not_implemented(
-                        "resource.getrlimit requires host_env feature",
-                    ))
-                }
-            },
-            1,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "setrlimit",
-        crate::make_builtin_function_with_arity(
-            "setrlimit",
-            |args| {
-                #[cfg(all(unix, feature = "host_env"))]
-                {
-                    if args.len() < 2 {
-                        return Err(crate::PyError::type_error(
-                            "setrlimit() requires 2 arguments",
-                        ));
-                    }
-                    let res = unsafe {
-                        if !pyre_object::is_int(args[0]) {
-                            return Err(crate::PyError::type_error(
-                                "setrlimit(): resource should be an integer",
-                            ));
-                        }
-                        pyre_object::w_int_get_value(args[0]) as libc::rlim_t
-                    };
-                    // limits is a 2-tuple (soft, hard).
-                    let (soft, hard) = unsafe {
-                        if !pyre_object::is_tuple(args[1]) || pyre_object::w_tuple_len(args[1]) != 2
-                        {
-                            return Err(crate::PyError::type_error(
-                                "setrlimit(): limits should be a tuple of (soft, hard)",
-                            ));
-                        }
-                        let s = pyre_object::w_tuple_getitem(args[1], 0).unwrap();
-                        let h = pyre_object::w_tuple_getitem(args[1], 1).unwrap();
-                        if !pyre_object::is_int(s) || !pyre_object::is_int(h) {
-                            return Err(crate::PyError::type_error(
-                                "setrlimit(): limits members must be integers",
-                            ));
-                        }
-                        (
-                            pyre_object::w_int_get_value(s) as libc::rlim_t,
-                            pyre_object::w_int_get_value(h) as libc::rlim_t,
-                        )
-                    };
-                    let rl = libc::rlimit {
-                        rlim_cur: soft,
-                        rlim_max: hard,
-                    };
-                    match rustpython_host_env::resource::setrlimit(res, rl) {
-                        Ok(()) => return Ok(pyre_object::w_none()),
-                        Err(e) => {
-                            // `lib_pypy/resource.py:89-95` — EINVAL and
-                            // EPERM both surface as ValueError with
-                            // distinct messages; all other errnos stay
-                            // as OSError.
-                            let errno = e.raw_os_error().unwrap_or(0);
-                            if errno == libc::EINVAL {
-                                return Err(crate::PyError::value_error(
-                                    "current limit exceeds maximum limit",
-                                ));
-                            }
-                            if errno == libc::EPERM {
-                                return Err(crate::PyError::value_error(
-                                    "not allowed to raise maximum limit",
-                                ));
-                            }
-                            return Err(crate::PyError::os_error_with_errno(
-                                errno,
-                                format!("setrlimit: {e}"),
-                            ));
-                        }
-                    }
-                }
-                #[cfg(not(all(unix, feature = "host_env")))]
-                {
-                    let _ = args;
-                    Err(crate::PyError::not_implemented(
-                        "resource.setrlimit requires host_env feature",
-                    ))
-                }
-            },
-            2,
-        ),
-    );
-    // ── Constants (POSIX subset matching CPython) ──
-    #[cfg(unix)]
-    {
-        crate::dict_storage_store(
-            ns,
-            "RUSAGE_SELF",
-            pyre_object::w_int_new(libc::RUSAGE_SELF as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "RUSAGE_CHILDREN",
-            pyre_object::w_int_new(libc::RUSAGE_CHILDREN as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "RLIMIT_CPU",
-            pyre_object::w_int_new(libc::RLIMIT_CPU as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "RLIMIT_FSIZE",
-            pyre_object::w_int_new(libc::RLIMIT_FSIZE as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "RLIMIT_DATA",
-            pyre_object::w_int_new(libc::RLIMIT_DATA as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "RLIMIT_STACK",
-            pyre_object::w_int_new(libc::RLIMIT_STACK as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "RLIMIT_CORE",
-            pyre_object::w_int_new(libc::RLIMIT_CORE as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "RLIMIT_NOFILE",
-            pyre_object::w_int_new(libc::RLIMIT_NOFILE as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "RLIMIT_AS",
-            pyre_object::w_int_new(libc::RLIMIT_AS as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "RLIMIT_RSS",
-            pyre_object::w_int_new(libc::RLIMIT_RSS as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "RLIMIT_NPROC",
-            pyre_object::w_int_new(libc::RLIMIT_NPROC as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "RLIMIT_MEMLOCK",
-            pyre_object::w_int_new(libc::RLIMIT_MEMLOCK as i64),
-        );
-        // RLIM_INFINITY: unsigned max — pyre stores as i64 (-1 on signed widen).
-        crate::dict_storage_store(
-            ns,
-            "RLIM_INFINITY",
-            pyre_object::w_int_new(libc::RLIM_INFINITY as i64),
-        );
-    }
 }
 
 /// itertools stub
