+++ pyre/pyre-interpreter/src/module/grp/moduledef.rs
@@ -1,0 +1,7 @@
+//! grp module definition — PyPy: lib_pypy/grp.py (via cffi).
+
+use crate::DictStorage;
+
+pub fn init(ns: &mut DictStorage) {
+    super::interp_grp::register_module(ns);
+}
