+++ pyre/pyre-interpreter/src/module/mod.rs
@@ -20,9 +20,12 @@ pub mod _weakref;
 pub mod atexit;
 pub mod faulthandler;
 pub mod fcntl;
+pub mod grp;
 pub mod math;
 pub mod mmap;
 pub mod operator;
+pub mod pwd;
+pub mod resource;
 pub mod select;
 pub mod sys;
 pub mod syslog;
