+++ pyre/pyre-interpreter/src/module/resource/mod.rs
@@ -1,0 +1,9 @@
+//! resource module — `lib_pypy/resource.py` (PyPy keeps it app-level via
+//! `_resource_cffi`); pyre takes Modules/resource.c shape since pyre has
+//! no app-level stdlib.
+//!
+//! Exposes getrusage / getrlimit / setrlimit plus the standard RUSAGE_*
+//! / RLIMIT_* constants and `struct_rusage`.
+
+pub mod interp_resource;
+pub mod moduledef;
