+++ pyre/pyre-interpreter/src/module/resource/moduledef.rs
@@ -1,0 +1,7 @@
+//! resource module definition — PyPy: lib_pypy/resource.py (via cffi).
+
+use crate::DictStorage;
+
+pub fn init(ns: &mut DictStorage) {
+    super::interp_resource::register_module(ns);
+}
